refactor: reestruturar projeto com diretório pages e otimizar imports
