+++ sigo-frontend/src/components/Modal.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import './Modal.css';
 
 function Modal({ ocorrencia, onClose }) {
