+++ sigo-frontend/src/index.css
@@ -1,6 +1,1 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@400;700&display=swap');
-
-body {
-  margin: 0;
-  font-family: 'Poppins', sans-serif;
-}
+/* Estilos globais movidos para App.css */
