+++ sigo-frontend/src/pages/Dashboard.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import { useState, useEffect } from 'react';
 import { FaFire, FaClock, FaRegClock, FaUsers } from 'react-icons/fa';
 import { FiFilePlus, FiClipboard, FiFileText, FiSettings, FiArrowUp } from 'react-icons/fi';
 import { IoDocumentTextOutline } from "react-icons/io5";
