+++ sigo-frontend/src/pages/Register.css
