+++ sigo-frontend/src/pages/Register.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
 import './Register.css';
 import logoSigo from '../assets/logosigo.svg';
