+++ sigo-frontend/src/pages/VerifyPage.css
